We try to:
* use large chunk size
* copy memory as much as possible

to make good use of the the RVV advance.

We get a ~5.4% performance gain in decompression when compared to load/store single 8-byte chunk on the SiFive internal FPGA.

Benchmark tool: [deflatebench](https://github.com/zlib-ng/deflatebench)