@KungFuJesus 2023-09-20
> I suspect that the RVV implementation of `CHUNKCOPY` would benefit other architectures that aren't using SIMD, especially if unaligned memory access is unavailable (such as on older ARM platforms).
> 
> It might be worth looking at `chunkcopy_safe` as well, since RISC-V doesn't seem to inline `memcpy` calls for more than 8 bytes.

Again, CHUNKCOPY_SAFE kinda sorta already does this:
https://github.com/zlib-ng/zlib-ng/blob/develop/inflate_p.h#L151