@Dead2 2023-09-23
I tested adding just the two if blocks to the standard chunkcopy, not sure whether anything else would be needed.
Decompression speeds:
C path (chunksize 8), inflate went from ~625ms to ~887ms.
AVX2 path (chunksize 32), inflate went from ~451ms to ~748ms.

So unless there is something to magically speed that up massively, this does not work in the standard chunkcopy implementation and should indeed be a RVV-specific implementation (I am amazed that this does speed things up there though, I assume you did verify that just upping the chunksize didn't account for the speedup)