@alexsifivetw 2023-09-24
We used `uint64_t` for chunk, implemented `chunkmemset_2`/`chunkmemset_4`/`chunkmemset_8`, and get a  `10.1%` (> `5.4%`) performance gain.