@ccawley2011 2023-09-25
> I tested adding just the two if blocks to the standard chunkcopy, not sure whether anything else would be needed.
> 
> Decompression speeds:
> 
> C path (chunksize 8), inflate went from ~625ms to ~887ms.
> 
> AVX2 path (chunksize 32), inflate went from ~451ms to ~748ms.
> 
> 
> 
> So unless there is something to magically speed that up massively, this does not work in the standard chunkcopy implementation and should indeed be a RVV-specific implementation (I am amazed that this does speed things up there though, I assume you did verify that just upping the chunksize didn't account for the speedup)

I suspect that what makes the difference here is probably one of two things:
- If loadchunk/storechunk aren't fully inlined, you end up with a worst case scenario of two memcpy calls for each chunk copied. However if it is inlined, you can avoid needing to call memcpy at all.
- If memcpy is repeatedly called with a small size, there's a chance that it'll spend most of the time taking care of alignment instead of doing larger loads and stores.

I haven't done any proper benchmarks though, so this is all just speculation.

It's likely that the discussion PR #1548 is related to a degree.